# chore(parser): Improve performance by inlining functions