--- a/bluejay-parser/src/ast/executable/executable_definition.rs
+++ b/bluejay-parser/src/ast/executable/executable_definition.rs
@@ -8,6 +8,7 @@ pub enum ExecutableDefinition<'a> {
 }
 
 impl<'a> FromTokens<'a> for ExecutableDefinition<'a> {
+    #[inline]
     fn from_tokens(tokens: &mut impl Tokens<'a>) -> Result<Self, ParseError> {
         if OperationDefinition::is_match(tokens) {
             OperationDefinition::from_tokens(tokens).map(Self::Operation)
@@ -20,6 +21,7 @@ impl<'a> FromTokens<'a> for ExecutableDefinition<'a> {
 }
 
 impl<'a> IsMatch<'a> for ExecutableDefinition<'a> {
+    #[inline]
     fn is_match(tokens: &mut impl Tokens<'a>) -> bool {
         OperationDefinition::is_match(tokens) || FragmentDefinition::is_match(tokens)
     }