--- a/bluejay-parser/src/ast/tokens.rs
+++ b/bluejay-parser/src/ast/tokens.rs
@@ -34,6 +34,7 @@ pub struct LexerTokens<'a, T: Lexer<'a>> {
 }
 
 impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
+    #[inline]
     pub fn new(lexer: T) -> Self {
         Self {
             lexer,
@@ -42,11 +43,13 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn peek<'b, 'c: 'b>(&'c mut self, idx: usize) -> Option<&'b LexicalToken> {
         self.compute_up_to(idx);
         self.buffer.get(idx)
     }
 
+    #[inline]
     pub fn peek_next(&mut self) -> Option<&LexicalToken> {
         self.peek(0)
     }
@@ -63,6 +66,7 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn expect_name(&mut self) -> Result<Name<'a>, ParseError> {
         match self.next() {
             Some(LexicalToken::Name(n)) => Ok(n),
@@ -71,6 +75,7 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn expect_variable(&mut self) -> Result<Variable<'a>, ParseError> {
         match self.next() {
             Some(LexicalToken::VariableName(n)) => Ok(n),
@@ -82,6 +87,7 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn expect_name_value(&mut self, value: &str) -> Result<Span, ParseError> {
         match self.next() {
             Some(LexicalToken::Name(n)) if n.as_str() == value => Ok(n.into()),
@@ -93,6 +99,7 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn expect_punctuator(
         &mut self,
         punctuator_type: PunctuatorType,
@@ -107,18 +114,21 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn unexpected_eof(&self) -> ParseError {
         ParseError::UnexpectedEOF {
             span: self.lexer.empty_span(),
         }
     }
 
+    #[inline]
     pub fn unexpected_token(&mut self) -> ParseError {
         self.next()
             .map(|token| ParseError::UnexpectedToken { span: token.into() })
             .unwrap_or_else(|| self.unexpected_eof())
     }
 
+    #[inline]
     fn next_if<F>(&mut self, f: F) -> Option<Span>
     where
         F: Fn(&LexicalToken) -> bool,
@@ -133,50 +143,61 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
         }
     }
 
+    #[inline]
     pub fn next_if_punctuator(&mut self, punctuator_type: PunctuatorType) -> Option<Span> {
         self.next_if(|t| matches!(t, LexicalToken::Punctuator(p) if p.r#type() == punctuator_type))
     }
 
+    #[inline]
     pub fn next_if_int_value(&mut self) -> Option<IntValue> {
         matches!(self.peek_next(), Some(LexicalToken::IntValue(_)))
             .then(|| self.next().unwrap().into_int_value().unwrap())
     }
 
+    #[inline]
     pub fn next_if_float_value(&mut self) -> Option<FloatValue> {
         matches!(self.peek_next(), Some(LexicalToken::FloatValue(_)))
             .then(|| self.next().unwrap().into_float_value().unwrap())
     }
 
+    #[inline]
     pub fn next_if_string_value(&mut self) -> Option<StringValue<'a>> {
         matches!(self.peek_next(), Some(LexicalToken::StringValue(_)))
             .then(|| self.next().unwrap().into_string_value().unwrap())
     }
 
+    #[inline]
     pub fn next_if_name(&mut self) -> Option<Name<'a>> {
         matches!(self.peek_next(), Some(LexicalToken::Name(_)))
             .then(|| self.next().unwrap().into_name().unwrap())
     }
 
+    #[inline]
     pub fn next_if_name_matches(&mut self, name: &str) -> Option<Span> {
         self.next_if(|t| matches!(t, LexicalToken::Name(n) if n.as_str() == name))
     }
 
+    #[inline]
     pub fn peek_name(&mut self, n: usize) -> Option<&Name> {
         self.peek(n).and_then(LexicalToken::as_name)
     }
 
+    #[inline]
     pub fn peek_variable_name(&mut self, n: usize) -> bool {
         matches!(self.peek(n), Some(LexicalToken::VariableName(_)))
     }
 
+    #[inline]
     pub fn peek_name_matches(&mut self, n: usize, name: &str) -> bool {
         matches!(self.peek_name(n), Some(n) if n.as_str() == name)
     }
 
+    #[inline]
     pub fn peek_string_value(&mut self, n: usize) -> bool {
         matches!(self.peek(n), Some(LexicalToken::StringValue(_)))
     }
 
+    #[inline]
     pub fn peek_punctuator_matches(&mut self, n: usize, punctuator_type: PunctuatorType) -> bool {
         matches!(self.peek(n), Some(LexicalToken::Punctuator(p)) if p.r#type() == punctuator_type)
     }
@@ -185,6 +206,7 @@ impl<'a, T: Lexer<'a>> LexerTokens<'a, T> {
 impl<'a, T: Lexer<'a>> Iterator for LexerTokens<'a, T> {
     type Item = LexicalToken<'a>;
 
+    #[inline]
     fn next(&mut self) -> Option<Self::Item> {
         self.compute_up_to(0);
         self.buffer.pop_front()
@@ -198,74 +220,92 @@ impl<'a, T: Lexer<'a>> From<LexerTokens<'a, T>> for Vec<(LexError, Span)> {
 }
 
 impl<'a, T: Lexer<'a>> Tokens<'a> for LexerTokens<'a, T> {
+    #[inline]
     fn expect_variable(&mut self) -> Result<Variable<'a>, ParseError> {
         self.expect_variable()
     }
 
+    #[inline]
     fn expect_name(&mut self) -> Result<Name<'a>, ParseError> {
         self.expect_name()
     }
 
+    #[inline]
     fn expect_name_value(&mut self, value: &str) -> Result<Span, ParseError> {
         self.expect_name_value(value)
     }
 
+    #[inline]
     fn expect_punctuator(&mut self, punctuator_type: PunctuatorType) -> Result<Span, ParseError> {
         self.expect_punctuator(punctuator_type)
     }
 
+    #[inline]
     fn unexpected_eof(&self) -> ParseError {
         self.unexpected_eof()
     }
 
+    #[inline]
     fn unexpected_token(&mut self) -> ParseError {
         self.unexpected_token()
     }
 
+    #[inline]
     fn next_if_punctuator(&mut self, punctuator_type: PunctuatorType) -> Option<Span> {
         self.next_if_punctuator(punctuator_type)
     }
 
+    #[inline]
     fn next_if_int_value(&mut self) -> Option<IntValue> {
         self.next_if_int_value()
     }
 
+    #[inline]
     fn next_if_float_value(&mut self) -> Option<FloatValue> {
         self.next_if_float_value()
     }
 
+    #[inline]
     fn next_if_string_value(&mut self) -> Option<StringValue<'a>> {
         self.next_if_string_value()
     }
 
+    #[inline]
     fn next_if_name(&mut self) -> Option<Name<'a>> {
         self.next_if_name()
     }
 
+    #[inline]
     fn next_if_name_matches(&mut self, name: &str) -> Option<Span> {
         self.next_if_name_matches(name)
     }
 
+    #[inline]
     fn peek_name(&mut self, n: usize) -> Option<&Name> {
         self.peek_name(n)
     }
 
+    #[inline]
     fn peek_variable_name(&mut self, n: usize) -> bool {
         self.peek_variable_name(n)
     }
 
+    #[inline]
     fn peek_name_matches(&mut self, n: usize, name: &str) -> bool {
         self.peek_name_matches(n, name)
     }
 
+    #[inline]
     fn peek_string_value(&mut self, n: usize) -> bool {
         self.peek_string_value(n)
     }
 
+    #[inline]
     fn peek_punctuator_matches(&mut self, n: usize, punctuator_type: PunctuatorType) -> bool {
         self.peek_punctuator_matches(n, punctuator_type)
     }
 
+    #[inline]
     fn into_errors(self) -> Vec<(LexError, Span)> {
         self.errors
     }