The rules applied for something to be marked as inlined is that either a small function or a function with as ingle callsite.

```
parse github schema definitions
                        time:   [3.7500 ms 3.7602 ms 3.7717 ms]
                        change: [-3.6874% -3.3392% -2.9969%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking parse kitchen sink executable document: Collecting 100 samples in estimate
parse kitchen sink executable document
                        time:   [10.215 µs 10.240 µs 10.272 µs]
                        change: [-11.600% -11.267% -10.949%] (p = 0.00 < 0.05)
                        Performance has improved.
```